+++ protocol/v2/qbft/config.go
@@ -1,4 +1,4 @@
-package types
+package qbft
 
 import (
 	specqbft "github.com/bloxapp/ssv-spec/qbft"
